+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 [v3.6.0](https://github.com/sedlatschek/werkbank3/releases/tag/v3.6.0.0)
 - fix bug where symlinks could not be deleted
+- add web button to open git remote links
 
 [v3.5.0](https://github.com/sedlatschek/werkbank3/releases/tag/v3.5.0.0)
 - fix bug where werk would disappear after a transition, if Werkbank was restarted after the delete operation was done